Moved sorting of sectors into sectors function
The sectors function previously just added Sector objects as it identified
them, meaning that it would put normal sectors before standbys etc. The sorting
was then done in the duties function. It is more useful to have the output of
sectors sorted, and the duties function can just assert that this is the case.