Move crew list from Duty structure to Sector structure
The crew list was originally in the Duty structure as this was how the vertical
roster presented it. The horizontal roster, however, gives crew lists for
individual sectors. This change prepares for using horizontal PDF rosters
should that become necessary, as well as it just making more sense in terms or
structure.