Restructure to improve modularity
The two processing modules, roster.py and logbook_report.py were using a
BeautifulSoup soup as the input to their main exported function. This assumes
that all inputs will be HTML. For future proofing, these functions are now
given the HTML itself and are completely responsible for processing it. This
will allow a future module to, for example, take a PDF as input and maintain
the same function signature for their main exported function.