Extract duty times from sectors when there are multiple sectors
It was previously assumed that if a duty did not have report and off duty
times then the duty would consist of a single item in the sectors field.
My Jan 2025 roster showed this assumption to be erroneous, so we now take
the duty times from the list of sectors.